[feat] fin seccion 13

+++ src/app/api/v1/todos/route.ts
@@ -2,6 +2,7 @@ import { NextResponse, NextRequest } from 'next/server';
 
 import prisma from '@/lib/prisma';
 import { boolean, object, string } from 'yup';
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
@@ -49,9 +50,16 @@ const postSchema = object({
 });
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json('Not authorized', { status: 401 });
+  }
+
   try {
     const { complete, description } = await postSchema.validate(await request.json());
-    const todo = await prisma.todo.create({ data: { description, complete } });
+    const todo = await prisma.todo.create({
+      data: { description, complete, userId: user.id },
+    });
 
     return NextResponse.json(
       {
@@ -68,10 +76,16 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json('Not authorized', { status: 401 });
+  }
+
   try {
     await prisma.todo.deleteMany({
       where: {
         complete: true,
+        userId: user.id,
       },
     });
